Add files via upload
In this project we have analyzed the superstore data from different regions of USA with different cities in it. Which region has the highest sales?  Profit and loss analysis in  each region and different categories . It answers to such relevant questions by providing descriptive, qualitative and quantitative analysis of superstore data.